Move CommentList to components
