+++ app/components/CommentList/index.js
@@ -1,0 +1,21 @@
+import React, { Component } from 'react'
+import Comment from '../../components/Comment'
+
+class CommentList extends Component {
+  render() {
+    const commentNodes = this.props.data.map((comment) => {
+      return (
+        <Comment author={comment.author} key={comment.id}>
+          {comment.text}
+        </Comment>
+      )
+    })
+    return (
+      <div className="commentList">
+        {commentNodes}
+      </div>
+    )
+  }
+}
+
+export default CommentList
